% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karpi-package.R
\docType{package}
\name{karpi-package}
\alias{karpi}
\alias{karpi-package}
\title{karpi: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
