% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_asset.R
\name{kpi_asset_info}
\alias{kpi_asset_info}
\title{Get survey asset metadata}
\usage{
kpi_asset_info(asset_id)
}
\arguments{
\item{asset_id}{The survey's asset ID (found after the "/forms/" in the URL)}
}
\value{
A list based on the JSON structure of the asset metadata
}
\description{
Fetches the overall asset metadata for a survey with
asset ID \code{asset_id}
}
\examples{
\dontrun{
  # Form url is this: https://kf.kobotoolbox.org/#/forms/abcdefghijklmnop
  # Asset ID is: abcdefghijklmnop
  karpi::kpi_asset_info("abcdefghijklmnop")
}
}
