% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_auth.R
\name{kpi_auth_token}
\alias{kpi_auth_token}
\title{Fetch or set KPI Auth Token}
\usage{
kpi_auth_token(token = NULL)
}
\arguments{
\item{token}{If not \code{NULL}, the value with which to overwrite the
"KPI_TOKEN" environment variable.}
}
\value{
The value of "KPI_TOKEN" or \code{NULL} if not defined
}
\description{
By default, this function reads from the "KPI_TOKEN" environment
variable. If \code{token} is provided, then the "KPI_TOKEN" environment
variable will be overwriten with the value of \code{token}.
}
