% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_data.R
\name{kpi_get_data}
\alias{kpi_get_data}
\title{Download a dataset}
\usage{
kpi_get_data(
  asset_id,
  expand_multiple = TRUE,
  expand_delimiter = ".",
  system_vars = "_uuid"
)
}
\arguments{
\item{asset_id}{The survey's asset ID (found after the "/forms/" in the URL)}

\item{expand_multiple}{If true, each choice in a multiple choice
field will get its own column. Expanding uses one-hot encoding.}

\item{expand_delimiter}{The expanded columns follow the pattern
"[field name][delimiter][choice value]". By default, the delimiter
is "." in karpi. Other ODK implementations use "/" as a delimiter.}

\item{system_vars}{Which ODK system variables to keep in the downloaded
dataset, e.g. "_uuid", "_id", and "_version". "_geolocation" is omitted
for privacy reasons.}
}
\value{
A data.frame of the dataset stored as \code{asset_id} on KoBo
}
\description{
Downloads a survey dataset with ID \code{asset_id} from KoBo.
}
\examples{
\dontrun{
  # Form url is this: https://kf.kobotoolbox.org/#/forms/abcdefghijklmnop
  # Asset ID is: abcdefghijklmnop
  karpi::kpi_get_data("abcdefghijklmnop")
}
}
