% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_data.R
\name{kpi_get_raw_data}
\alias{kpi_get_raw_data}
\title{Download a dataset in its JSON form}
\usage{
kpi_get_raw_data(asset_id)
}
\arguments{
\item{asset_id}{The survey's asset ID (found after the "/forms/" in the URL)}
}
\value{
A list of the dataset stored as \code{asset_id} on KoBo
}
\description{
Downloads a survey dataset with ID \code{asset_id} from KoBo, leaving the data
in its list-recursive format. This routine returns the data as-is: no
standard data processing occurs by default.
}
\examples{
\dontrun{
# Form url is this: https://kf.kobotoolbox.org/#/forms/abcdefghijklmnop
# Asset ID is: abcdefghijklmnop
karpi::kpi_get_raw_data("abcdefghijklmnop")
}
}
