% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_xlsform.R
\name{kpi_get_xlsform}
\alias{kpi_get_xlsform}
\title{Fetch the XLSForm for a survey with asset ID \code{asset_id}}
\usage{
kpi_get_xlsform(asset_id, file = NULL, verbose = TRUE)
}
\arguments{
\item{asset_id}{The survey's asset ID (found after the "/forms/" in the URL)}

\item{file}{If not \code{NULL}, the path where the XLSForm should be saved}

\item{verbose}{If \code{TRUE}, echos a message where the file was saved}
}
\value{
An \code{odk_xlsform} object, a \code{list()} with \code{id}, \code{survey},
\code{choices}, and \code{settings} entries.
}
\description{
Downloads the current XLSForm for a survey with ID \code{asset_id}.
}
\examples{
\dontrun{
  # Form url is this: https://kf.kobotoolbox.org/#/forms/abcdefghijklmnop
  # Asset ID is: abcdefghijklmnop
  form <- karpi::kpi_get_xlsform("abcdefghijklmnop")
  print(form$survey)
  print(form$choices)
}
}
