% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_url.R
\name{kpi_set_url}
\alias{kpi_set_url}
\title{Set KPI host domain}
\usage{
kpi_set_url(type = "kobotoolbox", domain = NULL)
}
\arguments{
\item{type}{Which KPI platform to use - one of "kobotoolbox", "humanitarian", or "custom"}

\item{domain}{Domain URL for "custom" type platform}
}
\value{
The KPI domain URL
}
\description{
karpi supports the kobotoolbox (kf) and humanitarian (kc) domains
by default. To point to another KPI domain, set \code{type} to "custom"
and set \code{domain} to the domain URL. Setting the URL overwrites the
"kpi.urlroot" option.
}
\section{Default KPI domains}{
KoBo has two domains for use:
\enumerate{
\item kobotoolbox (kf.kobotoolbox.org): General purpose KoBo deployments on KoBo
\item humanitarian (kc.kobotoolbox.info): Unrestricted KoBo platform used by humanitarian agencies (e.g. UNICEF)
}
}

