% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_url.R
\name{kpi_url}
\alias{kpi_url}
\title{Fetch current KPI domain URL}
\usage{
kpi_url()
}
\value{
The KPI domain URL
}
\description{
Reads from the "kpi.urlroot" option. If the value is \code{NULL},
an error is raised to notify the user to set the URL with
\code{kpi_set_url()}.
}
