% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_data.R
\name{kpi_get_data}
\alias{kpi_get_data}
\title{Download a dataset}
\usage{
kpi_get_data(
  asset_id,
  expand_multiple = TRUE,
  expand_delimiter = ".",
  drop_expanded = FALSE,
  include_geolocation = FALSE,
  system_vars = "_uuid",
  missing_vars = c("nothing", "warn", "msg", "err")
)
}
\arguments{
\item{asset_id}{The survey's asset ID (found after the "/forms/" in the URL)}

\item{expand_multiple}{If true, each choice in a multiple choice
field will get its own column. Expanding uses one-hot encoding.}

\item{expand_delimiter}{The expanded columns follow the pattern
"[field name][delimiter][choice value]". By default, the delimiter
is "." in karpi. Other ODK implementations use "/" as a delimiter.}

\item{drop_expanded}{Drops multiple choices columns that were expanded into
dummies.}

\item{include_geolocation}{If \code{TRUE}, keeps "_geolocation" in data. Defaults
to \code{FALSE}.}

\item{system_vars}{Which ODK system variables to keep in the downloaded
dataset, e.g. "_uuid", "_id", and "_version". "_geolocation" is omitted
for privacy reasons by default.}

\item{missing_vars}{What to do if variables expected in the KoBo XForm meta
are missing. Defaults to nothing. Usually this is the result of empty
or unused fields during data collection.}
}
\value{
A data.frame of the dataset stored as \code{asset_id} on KoBo
}
\description{
Downloads a survey dataset with ID \code{asset_id} from KoBo.
}
\examples{
\dontrun{
# Form url is this: https://kf.kobotoolbox.org/#/forms/abcdefghijklmnop
# Asset ID is: abcdefghijklmnop
karpi::kpi_get_data("abcdefghijklmnop")
}
}
